optimize for hash join

Created-by: liuyu 00809192
Author-id: 53630
MR-id: 15831170
Commit-by: liuyu;Shunning Jiang
Merged-by: liaodenghong 00447528
E2E-issues: issue430
Description: + [ ] I've added/changed the documentation
+ [ ] I've added test for the code

See merge request: Kunpeng-Computing-DC/BigData/OmniOperatorJIT!506

+++ core/src/operator/aggregation/group_aggregation.cpp
@@ -192,8 +192,8 @@ OmniStatus HashAggregationOperator::Init()
     }
 
     // agg source types
-    for (int i = 0; i < aggInputCols.size(); ++i) {
-        for (int j = 0; j < aggInputCols[i].size(); ++j) {
+    for (size_t i = 0; i < aggInputCols.size(); ++i) {
+        for (size_t j = 0; j < aggInputCols[i].size(); ++j) {
             sourceTypes[aggInputCols[i][j]] = aggInputTypes[i].GetType(j)->GetId();
         }
     }
@@ -262,7 +262,7 @@ int32_t HashAggregationOperator::InitMaxRowCountAndOutputTypes()
 void HashAggregationOperator::SetVectors(mem::Allocator *vecAllocator, VectorBatch *vectorBatch,
     const std::vector<DataTypePtr> &types, int32_t rowCount)
 {
-    for (int colIndex = 0; colIndex < types.size(); ++colIndex) {
+    for (size_t colIndex = 0; colIndex < types.size(); ++colIndex) {
         const DataTypePtr &type = types[colIndex];
         GROUP_AGG_FUNCTIONS[type->GetId()].setVector(vectorBatch, *type, colIndex, vecAllocator, rowCount);
     }
